+++ src/components/ui/BaseLoader.vue
@@ -6,12 +6,15 @@
       class="fa-spin-pulse spinner-icon"
       :class="{ 'cover-page': props.coverPage }"
       size="6x"
+      :color="shouldChangeColor ? 'white' : 'black'"
     />
     <slot v-else></slot>
   </transition>
 </template>
 
 <script>
+import store from '@/lib/store';
+
 export default {
   name: 'BaseLoader',
   props: {
@@ -26,6 +29,18 @@ export default {
     coverPage: {
       type: Boolean,
       default: false,
+    },
+    enableDarkmodeColorSwitch: {
+      type: Boolean,
+      default: true,
+    },
+  },
+  computed: {
+    isDarkModeEnabled() {
+      return store.state.isDarkModeEnabled;
+    },
+    shouldChangeColor() {
+      return this.isDarkModeEnabled && this.enableDarkmodeColorSwitch;
     },
   },
 };