+++ src/views/PokemonsView.vue
@@ -1,22 +1,30 @@
 <template>
-  <CenteredColumn>
+  <div class="pokemons-view">
     <ErrorBoundary
       componentName="PokemonList"
       errorMessage="Unable to load Pokemon list"
     >
       <PokemonList />
     </ErrorBoundary>
-  </CenteredColumn>
+  </div>
 </template>
 
 <script>
-import CenteredColumn from '@/components/ui/CenteredColumn.vue';
 import ErrorBoundary from '@/components/ui/ErrorBoundary.vue';
 import PokemonList from '@/components/pokemons/PokemonList.vue';
 
 export default {
   name: 'PokemonsView',
   title: 'Pokemons',
-  components: { CenteredColumn, PokemonList, ErrorBoundary },
+  components: { PokemonList, ErrorBoundary },
 };
 </script>
+
+<style lang="scss" scoped>
+.pokemons-view {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+}
+</style>